Replace winapi dependency with windows; Better error messages
The windows crate has improved since commit 6e929c5 which replaced this
crate's dependency on it with the winapi crate. While the issue of
missing trait documentations still exists, it seems that a PCWSTR can
now be constructed as a newtype from the same raw pointer type that was
used by winapi, and error handling has been improved considerably.